BUG-26554 Windows MSI changes for use of Oozie + Tez jobs with hdp.version

+++ winpkg/src/Scripts/InstallApi.psm1
@@ -255,8 +255,15 @@ function Configure(
         ### Copying hive-site and tez-site xmls to etc\action-conf\hive
         ###
         Write-Log "Copying hive-site and tez-site xmls to etc\action-conf\hive"
+        if (-not (Test-Path "$ENV:OOZIE_HOME\conf\action-conf\hive"))
+        {
+            New-Item -Path "$ENV:OOZIE_HOME\conf\action-conf\hive" -ItemType Directory -Force -ErrorAction Stop
+        }
         Copy-Item -Path "$ENV:HIVE_HOME\conf\hive-site.xml" -Destination "$ENV:OOZIE_HOME\conf\action-conf\hive\hive-site.xml" -Force -ErrorAction Stop
-        Copy-Item -Path "$ENV:TEZ_HOME\conf\tez-site.xml" -Destination "$ENV:OOZIE_HOME\conf\action-conf\hive\tez-site.xml" -Force -ErrorAction Stop
+        if (Test-Path "$ENV:TEZ_HOME\conf\tez-site.xml")
+        {
+            Copy-Item -Path "$ENV:TEZ_HOME\conf\tez-site.xml" -Destination "$ENV:OOZIE_HOME\conf\action-conf\hive\tez-site.xml" -Force -ErrorAction Stop
+        }
     }
     else
     {
